Update Arena commands to be more user friendly
The !arena command will now start as not joinable by default.

The !arena @someone will also @them in the response which is super
useful for sending the arena to someone and not have it get lost in
chat.

The !editarena nj command will now auto send the arena is not joinable in
twitch chat. This will also help viewers understand when the arena is
joinable and not joinable